stats/requirements.txt
psycopg2-binary
redis
numpy